cleaned up duplicate method error. subsequently dropped reek smells by a significant amount after a lot of refactoring. Will need to adjust code on view pages, etc, as well as methods used to test. Will not push main branch to master until tests are passing and everything is functional again

+++ app/models/room.rb
@@ -1,28 +1,45 @@
 class Room < ActiveRecord::Base
   mount_uploader :room_image, MainUploader
 
+  def initialize 
+    @items_collection = Item.where({"room_id" => self.id})
+    @item_id_array = []
+    @bad_items = {}
+    @good_items = {}
+    @creator = nil
+    @editor = nil
+  end
+
+  #RETURNS nil OR @items_collection instance variable
+  def get_find_items_for_room
+    return @items_collection
+  end
+
+  # Method finds all Item Object instances where "room_id" matches the current Room Instance ID && stores the item id in @item_id_array instance variable.
+  # SETS Array of Item Collection IDs to Instance Variable @item_id_array
+  def set_items_id_array_for_room
+    @item_id_array = Item.where({"room_id" => self.id}).pluck(:id)
+  end
+
+
+  #RETURNS nil OR @item_id_array instance variable
+  def get_items_id_array_for_room
+    return @item_id_array
+  end
 
   # Method searches all Items for objects associated with the Room, 
   # orders those Items by their "condition" value,
   # and defines the "bad condition" parameter with an array of integers.
   # Items whose condition are included in the array are pushed into an
   # array ("@bad_items").
-  #
-  # RETURNS an Array of Item objects, if not empty (an empty array will return # a nil value).  
-  def find_damaged_items_for_room
-    items = Item.where({"room_id" => self.id}) == nil ? nil : Item.where({"room_id" => self.id}).order('condition')
-    bad_condition = [1, 2, 3]
-    @bad_items = []
+  def set_find_damaged_items_for_room
+    @bad_items = Item.where({"room_id" => self.id, "condition" => 1..3}).order('condition')
+  end
 
-    if items != nil
-      items.each do |item|
-        item_condition = item.condition
-        if bad_condition.include?(item_condition)
-          @bad_items << item
-        end
-      end
-    end 
-    return @bad_items.empty? ? nil : @bad_items
+
+  # RETURNS an hash of Item objects or an empty hash
+  def get_find_damaged_items_for_room
+    return @bad_items
   end
 
   # Method searches all Items for objects associated with the Room, 
@@ -32,20 +49,12 @@ class Room < ActiveRecord::Base
   # array ("@good_items").
   #
   # RETURNS an Array of Item objects, if not empty (an empty array will return # a nil value).  
-  def find_good_items_for_room
-    items = Item.where({"room_id" => self.id}) == nil ? nil : Item.where({"room_id" => self.id}).order('condition desc')
-    good_condition = [4, 5]
-    @good_items = []
+  def set_find_good_items_for_room
+    @good_items = Item.where({"room_id" => self.id, "condition" => 4..5}).order('condition desc')
+  end
 
-    if items != nil
-      items.each do |item|
-        item_condition = item.condition
-        if good_condition.include?(item_condition)
-          @good_items << item
-        end
-      end
-    end 
-    return @good_items.empty? ? nil : @good_items
+  def get_find_good_items_for_room
+    return @good_items
   end
 
   # Method searches all Items for objects associated with the Room,
@@ -55,21 +64,12 @@ class Room < ActiveRecord::Base
   #
   # RETURNS nil
   def find_and_delete_item_photos_for_room
-    if Item.where({"room_id" => self.id}) != nil
-      item_arr = []
-      items = Item.where({"room_id" => self.id})
-
-      items.each do |item|
-        item_arr << item.id
-      end
-
-      item_arr. each do |item|
-        if Photo.where({"item_id" => item}) != nil
-          Photo.where({"item_id" => item}).delete_all
+      @item_id_array. each do |item|
+        item_photo = Photo.where({"item_id" => item}) 
+        if item_photo != nil
+          item_photo.delete_all
         end
       end
-
-    end
   end
 
   # Method searches all Items for objects associated with the Room,
@@ -77,25 +77,31 @@ class Room < ActiveRecord::Base
   #
   # RETURNS nil
   def find_and_delete_items_for_room
-    if Item.where({"room_id" => self.id}) != nil
-      Item.where({"room_id" => self.id}).delete_all
-    end
+     Item.where({"room_id" => self.id}).delete_all if @items_collection != nil
   end
 
   # Method RETURNS the User object associated with the "created_by"
   # attribute of the Room.
   #
   # RETURNS a User object, or nil if "created_by" has no value
+  def set_created_by_user_info_for_room
+    @creator = User.find_by_id(self.created_by)
+  end
+
   def get_created_by_user_info_for_room
-    return self.created_by == nil ?  nil : User.find_by_id(self.created_by)
+    return @creator
   end
 
   # Method RETURNS the User object associated with the "updated_by"
   # attribute of the Room.
   #
   # RETURNS a User object, or nil if "updated_by" has no value
+  def set_updated_by_user_info_for_room
+    @editor = User.find_by_id(self.updated_by)
+  end
+
   def get_updated_by_user_info_for_room
-    return self.updated_by == nil ?  nil : User.find_by_id(self.updated_by)
+    return @editor
   end
 
   # RETURNS Boolean if title value String is empty or not
